+++ src/reproject.h
@@ -25,7 +25,8 @@
  *	int * tarNNSouID:	the output IDs of nearest neighboring source cells 
  *	double * tarNNDis	the output nearest distance for each target cell (input NULL if you don't need this field)
  */ 
-void nearestNeighborBlockIndex(double ** psouLat, double ** psouLon, int nSou, double * tarLat, double * tarLon, int * tarNNSouID, double * tarNNDis, int nTar, double maxR);
+//void nearestNeighborBlockIndex(double ** psouLat, double ** psouLon, int nSou, double * tarLat, double * tarLon, int * tarNNSouID, double * tarNNDis, int nTar, double maxR);
+void nearestNeighborBlockIndex(double ** psouLat, double ** psouLon, uint64_t nSou, double * tarLat, double * tarLon, uint64_t * tarNNSouID, double * tarNNDis, uint64_t nTar, double maxR);
 
 
 /**
@@ -59,7 +60,8 @@ void nearestNeighbor(double ** psouLat, double ** psouLon, int nSou, double * ta
  * Output: 	
  * 	double * tarVal:	the output values at target cells
  */ 
-void nnInterpolate(double * souVal, double * tarVal, int * tarNNSouID, int nTar);
+//void nnInterpolate(double * souVal, double * tarVal, int * tarNNSouID, int nTar);
+void nnInterpolate(double * souVal, double * tarVal, uint64_t * tarNNSouID, uint64_t nTar);
 
 
 /**
@@ -78,7 +80,8 @@ void nnInterpolate(double * souVal, double * tarVal, int * tarNNSouID, int nTar)
  * 	double * tarSD:		the standard deviation (SD) value at target cells (can be NULL if no SD values need to be reported)
  * 	int * nSouPixels:	the output numbers of contributing source cells to each target cell
  */
-void summaryInterpolate(double * souVal, int * souNNTarID, int nSou, double * tarVal, double * tarSD, int * nSouPixels, int nTar);
+//void summaryInterpolate(double * souVal, int * souNNTarID, int nSou, double * tarVal, double * tarSD, int * nSouPixels, int nTar);
+void summaryInterpolate(double * souVal, uint64_t * souNNTarID, uint64_t nSou, double * tarVal, double * tarSD, uint64_t * nSouPixels, uint64_t nTar);
